Avoid transmitting private key over the network.
Previously the private key of a docker node was fetched over the web
interface to the ca-server, which is not encrypted. This changes the
shape of the handshake so that the docker node generates its own key
and a certificate signing request (CSR) which it then sends to the
ca-server, obtaining its signed certificate in the response to the
REST request.

This also adds some sensors, the idea having been that we would do
away with the web server entirely and use sensors to expose data and
an effector to sign the CSR.  It turns out that the effector call
will need a bit more work in brooklyn, so for now we retain the
web server interaction to transmit the certificate.

This also adds a script as a convenience for developers to get
certificates from the ca-server, getcert.sh. See the script for
comments on usage.